Split custom into submodule
