+++ src/oemof/solph/custom/__init__.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+from .electrical_line import ElectricalBus  # noqa: F401
+from .electrical_line import ElectricalLine  # noqa: F401
+from .generic_caes import GenericCAES  # noqa: F401
+from .link import Link  # noqa: F401
+from .piecewise_linear_transformer import (
+    PiecewiseLinearTransformer)  # noqa: F401
+from .sink_dsm import SinkDSM  # noqa: F401
